requirements.txt
# ------------------------------------
# Copyright (c) Pranjal Mishra.
# Licensed under the MIT License.
# Built Using Python 3.12.6
# ------------------------------------
fastapi==0.115.6
uvicorn==0.32.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.19
mongoengine==0.27.0
pymongo==3.12.0

